backend/requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==1.10.13  # ✅ Downgraded for LangChain 0.1.0 compatibility

# MongoDB
pymongo==4.6.1
motor==3.3.2

# AI & LLM
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-openai==0.0.5
langgraph==0.0.20

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2
kaleido==0.2.1

# HTTP & CORS
python-multipart==0.0.6
httpx==0.26.0

# Testing
colorama==0.4.6
requests==2.31.0
